server/requirements_minio.txt
# MinIO 相关依赖
minio==7.2.5
urllib3>=1.26.0
certifi
python-dotenv==1.0.0
